+++ compiler/lib/config.mli
@@ -84,6 +84,8 @@ module Flag : sig
 
   val wasi : unit -> bool
 
+  val split_toplevel : unit -> bool
+
   val enable : string -> unit
 
   val disable : string -> unit
@@ -104,6 +106,8 @@ module Param : sig
   val constant_max_depth : unit -> int
 
   val merge_node_max : unit -> int
+
+  val toplevel_split_size : unit -> int
 
   type tc =
     | TcNone